fix: portability and deployment improvements
- Fix include path in ext.h (sandbox.h -> sandbox/sandbox.h)
- Add log initialization in FastCGI bridge
- Fix %zu format specifier for portability in fcgi_io.c
- Guard SYS_poll/SYS_dup2 with #ifdef in sandbox.c
- Refactor Dockerfile: integrate nginx + SPA, use mirror sources
- Switch nginx upstream from Unix socket to TCP 9000
- Use unique session names with Date.now() in web UI
- Add .dockerignore to exclude build artifacts

+++ src/fastcgi/fcgi_io.c
@@ -52,10 +52,10 @@ void reply_json(request_t *r, int status, const char *body) {
 	FCGX_FPrintF(r->fcgx->out,
 		"Status: %s\r\n"
 		"Content-Type: application/json; charset=utf-8\r\n"
-		"Content-Length: %zu\r\n"
+		"Content-Length: %lu\r\n"
 		"Cache-Control: no-store\r\n"
 		"\r\n",
-		status_text(status), blen);
+		status_text(status), (unsigned long)blen);
 	if (blen) FCGX_PutStr(body, (int)blen, r->fcgx->out);
 }
